Stanza Security Exception done

+++ pojo-wanda/src/main/java/com/stanzaliving/wanda/venta/response/InstalmentPlanResponseDto.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 public class InstalmentPlanResponseDto {
     private String name;
     private InstallmentTypeEnum instalmentTypeEnum;
-    private Integer upFrontPaymentMonth;
+    private Double upFrontPaymentMonth;
 }
